Continuing the analysis
In this sequel I've improved the look of the annual sales chart by adding Title, the year in the labels, removing the side and top lines and putting the number of sales on top of each bar.

In addition, I continued studying the data and analyzed sales by month using the same logic as the analysis by year, and built a graph for visualization.